Added listeners for dependency hydration
So, loading dependencies in parallel means that they can step on each other if we're
not careful - who knew!

There are now three distinct phases of dependency hydration:

1. undefined: This is the default phase. It means the hydration system is not aware a
   dependency needs to be hydrated.
2. false: This phase indicates that the hydration system is aware of a dependency, but it
   is not hydrated yet.
3. true: This phase indicates that a dependency is completely hydrated and ready to go.

When a dependency is in the false phase, any subclasses trying to initialize are added to a list
and notified when the dependency moves to the true phase.

This forces subclasses to wait on their initialization until their superclass has all it's dependencies
fully hydrated.