Added tests, tried to demostrate async hydration/initialization
Unfortunately, I couldn't write a test that clearly showed the how initialization listeners
are added when maduls hydrate. I tried everything I could think of.

The ironic thing is that there's no good way to test this because of how complex and
nuanced the orchestration is. By design, async/await and Promises don't provide a way
to pause, or peek in on, behavior as it's happening. That's good pretty much all the
time; this just happens to be an edgiest-of-edge cases 😵