Fixed soem bugs, added some features
All the updates here came from working on jyn.

Boostrapper was updated to take two new arguments: params and config.

* params: These are params passed to the madul initializers and all dependent madul initializers.
* config: This allows you tp speficy a custom root and sdk for the bootstrapped madul and all dependent maduls.
    * root specifies the root of the file structure from which to load maduls. By default, it's process.cwd() but can be overridden
    sdk specifies the sdk functions to provide to the madul. This is merged with the value of the sdk property from any bundles including the madul

The order of operations for bootstrapping was udpated to accommodate the new arguments.

A fix was implemented for an issue encountered while working on jyn where threads were stomping on each other writing the contets of madul files. More work needs to be done, as now wrapped madul fiels are not refreshed/updated when they change. I'm thinking an env flag might be good to accommodate for this; something life MADUL_FORCE_RESAVE or something ...