madul methods are now, essentially, entirely functional
Instead of attaching deps to `this`, thery're now passed to methods as arguments.

Additionally, `null` is now being passed to `call`, so madul methods will no longer have any access to `this` whatsoever. The idea is that all madul methods will be arrow functions moving forward. I think this really simplifies things.

Since methods no longer have access to `this`, a new param, `self`, is being passed as an argument. This is one of my favorite things about Python ☺️